ignore the dependency-reduced-pom
The shade plugin creates an optimized pom for building the uber JAR.
This file is placed in the project basedir. Changing this location will
modify the basedir for all maven executions after shade, so we'll leave
it in the regular basedir and ignore it.

See http://maven.apache.org/plugins/maven-shade-plugin/shade-mojo.html
for details.